Move recording function into separate utilities file.
